+++ timeutil/timeutil.go
@@ -15,7 +15,7 @@ import (
 	"strings"
 	"time"
 
-	"pkg.re/essentialkaos/ek.v1/fmtutil"
+	"pkg.re/essentialkaos/ek.v2/fmtutil"
 )
 
 // ////////////////////////////////////////////////////////////////////////////////// //